Update suds and grpcio tests.  Unify spans in span.py
