+++ instana/span.py
@@ -1,5 +1,7 @@
-from basictracer.span import BasicSpan
 from .log import logger
+from .util import DictionaryOfStan
+from basictracer.span import BasicSpan
+import opentracing.ext.tags as ot_tags
 
 
 class InstanaSpan(BasicSpan):
@@ -10,7 +12,6 @@ class InstanaSpan(BasicSpan):
 
     def log_exception(self, e):
         try:
-            logger.debug("Logging error for span %s" % self.operation_name)
             message = ""
 
             self.set_tag("error", True)
@@ -24,10 +25,12 @@ class InstanaSpan(BasicSpan):
 
             if self.operation_name in ['rpc-server', 'rpc-client']:
                 self.set_tag('rpc.error', message)
-            elif self.operation_name == "postgres":
-                self.set_tag('pg.error', message)
             elif self.operation_name == "mysql":
                 self.set_tag('mysql.error', message)
+            elif self.operation_name == "postgres":
+                self.set_tag('pg.error', message)
+            elif self.operation_name == "soap":
+                self.set_tag('http.error', message)
             else:
                 self.log_kv({'message': message})
         except Exception:
@@ -57,3 +60,268 @@ class InstanaSpan(BasicSpan):
         return logs
 
 
+class BaseSpan(object):
+    def __str__(self):
+        return "BaseSpan(%s)" % self.__dict__.__str__()
+
+    def __repr__(self):
+        return self.__dict__.__str__()
+
+    def __init__(self, span, source, **kwargs):
+        self.t = span.context.trace_id
+        self.p = span.parent_id
+        self.s = span.context.span_id
+        self.ts = int(round(span.start_time * 1000))
+        self.d = int(round(span.duration * 1000))
+        self.f = source
+        self.ec = span.tags.pop("ec", None)
+        self.error = span.tags.pop("error", None)
+
+        if span.stack:
+            self.stack = span.stack
+
+        self.__dict__.update(kwargs)
+
+
+class SDKSpan(BaseSpan):
+    ENTRY_KIND = ["entry", "server", "consumer"]
+    EXIT_KIND = ["exit", "client", "producer"]
+
+    def __init__(self, span, source, service_name, **kwargs):
+        super(SDKSpan, self).__init__(span, source, **kwargs)
+        self.n = "sdk"
+        self.k = self.get_span_kind_as_int(span)
+
+        self.data = DictionaryOfStan()
+        self.data["sdk"]["name"] = span.operation_name
+        self.data["sdk"]["type"] = self.get_span_kind_as_string(span)
+        self.data["sdk"]["custom"]["tags"] = span.tags
+        self.data["sdk"]["custom"]["logs"] = span.logs
+        self.data["service"] = service_name
+
+        # self.data = Data()
+        # self.data.sdk = SDKData(name=span.operation_name, Type=self.get_span_kind_as_string(span))
+        # self.data.sdk.custom = CustomData(tags=span.tags, logs=span.collect_logs())
+        # self.data.service = service_name
+
+        if "arguments" in span.tags:
+            self.data.sdk.arguments = span.tags["arguments"]
+
+        if "return" in span.tags:
+            self.data.sdk.Return = span.tags["return"]
+
+        if len(span.context.baggage) > 0:
+            self.data["baggage"] = span.context.baggage
+
+    def get_span_kind_as_string(self, span):
+        """
+            Will retrieve the `span.kind` tag and return the appropriate string value for the Instana backend or
+            None if the tag is set to something we don't recognize.
+
+        :param span: The span to search for the `span.kind` tag
+        :return: String
+        """
+        kind = None
+        if "span.kind" in span.tags:
+            if span.tags["span.kind"] in self.ENTRY_KIND:
+                kind = "entry"
+            elif span.tags["span.kind"] in self.EXIT_KIND:
+                kind = "exit"
+            else:
+                kind = "intermediate"
+        return kind
+
+    def get_span_kind_as_int(self, span):
+        """
+            Will retrieve the `span.kind` tag and return the appropriate integer value for the Instana backend or
+            None if the tag is set to something we don't recognize.
+
+        :param span: The span to search for the `span.kind` tag
+        :return: Integer
+        """
+        kind = None
+        if "span.kind" in span.tags:
+            if span.tags["span.kind"] in self.ENTRY_KIND:
+                kind = 1
+            elif span.tags["span.kind"] in self.EXIT_KIND:
+                kind = 2
+            else:
+                kind = 3
+        return kind
+
+
+class RegisteredSpan(BaseSpan):
+    HTTP_SPANS = ("aiohttp-client", "aiohttp-server", "django", "http", "soap", "tornado-client",
+                  "tornado-server", "urllib3", "wsgi")
+
+    EXIT_SPANS = ("aiohttp-client", "cassandra", "couchbase", "log", "memcache", "mongo", "mysql", "postgres",
+                  "rabbitmq", "redis", "rpc-client", "sqlalchemy", "soap", "tornado-client", "urllib3",
+                  "pymongo")
+
+    ENTRY_SPANS = ("aiohttp-server", "aws.lambda.entry", "django", "wsgi", "rabbitmq", "rpc-server", "tornado-server")
+
+    LOCAL_SPANS = ("render")
+
+    def __init__(self, span, source, **kwargs):
+        super(RegisteredSpan, self).__init__(span, source, **kwargs)
+        self.n = span.operation_name
+        self.data = DictionaryOfStan()
+
+        self.k = 1
+        if span.operation_name in self.ENTRY_SPANS:
+            # entry
+            self._populate_entry_span_data(span)
+        elif span.operation_name in self.EXIT_SPANS:
+            self.k = 2 # exit
+            self._populate_exit_span_data(span)
+        elif span.operation_name in self.LOCAL_SPANS:
+            self.k = 3 # intermediate span
+            self._populate_local_span_data(span)
+
+        if "rabbitmq" in self.data and self.data["rabbitmq"]["sort"] == "consume":
+            self.k = 1  # entry
+
+        # Store any leftover tags in the custom section
+        if len(span.tags):
+            self.data["custom"]["tags"] = span.tags
+
+
+    def _populate_entry_span_data(self, span):
+        if span.operation_name in self.HTTP_SPANS:
+            self._collect_http_tags(span)
+        elif span.operation_name == "aws.lambda.entry":
+            self.data["lambda"]["arn"] = span.tags.pop('lambda.arn', "Unknown")
+            self.data["lambda"]["alias"] = None
+            self.data["lambda"]["runtime"] = "python"
+            self.data["lambda"]["functionName"] = span.tags.pop('lambda.name', "Unknown")
+            self.data["lambda"]["functionVersion"] = span.tags.pop('lambda.version', "Unknown")
+            self.data["lambda"]["error"] = None
+
+        elif span.operation_name == "rabbitmq":
+            self.data["rabbitmq"]["exchange"] = span.tags.pop('exchange', None)
+            self.data["rabbitmq"]["queue"] = span.tags.pop('queue', None)
+            self.data["rabbitmq"]["sort"] = span.tags.pop('sort', None)
+            self.data["rabbitmq"]["address"] = span.tags.pop('address', None)
+            self.data["rabbitmq"]["key"] = span.tags.pop('key', None)
+
+        elif span.operation_name == "rpc-server":
+            self.data["rpc"]["flavor"] = span.tags.pop('rpc.flavor', None)
+            self.data["rpc"]["host"] = span.tags.pop('rpc.host', None)
+            self.data["rpc"]["port"] = span.tags.pop('rpc.port', None)
+            self.data["rpc"]["call"] = span.tags.pop('rpc.call', None)
+            self.data["rpc"]["call_type"] = span.tags.pop('rpc.call_type', None)
+            self.data["rpc"]["params"] = span.tags.pop('rpc.params', None)
+            self.data["rpc"]["baggage"] = span.tags.pop('rpc.baggage', None)
+            self.data["rpc"]["error"] = span.tags.pop('rpc.error', None)
+        else:
+            logger.debug("SpanRecorder: Unknown entry span: %s" % span.operation_name)
+
+    def _populate_local_span_data(self, span):
+        if span.operation_name == "render":
+            self.data["render"]["name"] = span.tags.pop('name', None)
+            self.data["render"]["type"] = span.tags.pop('type', None)
+            self.data["log"]["message"] = span.tags.pop('message', None)
+            self.data["log"]["parameters"] = span.tags.pop('parameters', None)
+        else:
+            logger.debug("SpanRecorder: Unknown local span: %s" % span.operation_name)
+
+    def _populate_exit_span_data(self, span):
+        if span.operation_name in self.HTTP_SPANS:
+            self._collect_http_tags(span)
+        elif span.operation_name == "rabbitmq":
+            self.data["rabbitmq"]["exchange"] = span.tags.pop('exchange', None)
+            self.data["rabbitmq"]["queue"] = span.tags.pop('queue', None)
+            self.data["rabbitmq"]["sort"] = span.tags.pop('sort', None)
+            self.data["rabbitmq"]["address"] = span.tags.pop('address', None)
+            self.data["rabbitmq"]["key"] = span.tags.pop('key', None)
+
+        elif span.operation_name == "cassandra":
+            self.data["cassandra"]["cluster"] = span.tags.pop('cassandra.cluster', None)
+            self.data["cassandra"]["query"] = span.tags.pop('cassandra.query', None)
+            self.data["cassandra"]["keyspace"] = span.tags.pop('cassandra.keyspace', None)
+            self.data["cassandra"]["fetchSize"] = span.tags.pop('cassandra.fetchSize', None)
+            self.data["cassandra"]["achievedConsistency"] = span.tags.pop('cassandra.achievedConsistency', None)
+            self.data["cassandra"]["triedHosts"] = span.tags.pop('cassandra.triedHosts', None)
+            self.data["cassandra"]["fullyFetched"] = span.tags.pop('cassandra.fullyFetched', None)
+            self.data["cassandra"]["error"] = span.tags.pop('cassandra.error', None)
+
+        elif span.operation_name == "couchbase":
+            self.data["couchbase"]["hostname"] = span.tags.pop('couchbase.hostname', None)
+            self.data["couchbase"]["bucket"] = span.tags.pop('couchbase.bucket', None)
+            self.data["couchbase"]["type"] = span.tags.pop('couchbase.type', None)
+            self.data["couchbase"]["error"] = span.tags.pop('couchbase.error', None)
+            self.data["couchbase"]["error_type"] = span.tags.pop('couchbase.error_type', None)
+            self.data["couchbase"]["sql"] = span.tags.pop('couchbase.sql', None)
+
+        elif span.operation_name == "redis":
+            self.data["redis"]["connection"] = span.tags.pop('connection', None)
+            self.data["redis"]["driver"] = span.tags.pop('driver', None)
+            self.data["redis"]["command"] = span.tags.pop('command', None)
+            self.data["redis"]["error"] = span.tags.pop('redis.error', None)
+            self.data["redis"]["subCommands"] = span.tags.pop('subCommands', None)
+
+        elif span.operation_name == "rpc-client":
+            self.data["rpc"]["flavor"] = span.tags.pop('rpc.flavor', None)
+            self.data["rpc"]["host"] = span.tags.pop('rpc.host', None)
+            self.data["rpc"]["port"] = span.tags.pop('rpc.port', None)
+            self.data["rpc"]["call"] = span.tags.pop('rpc.call', None)
+            self.data["rpc"]["call_type"] = span.tags.pop('rpc.call_type', None)
+            self.data["rpc"]["params"] = span.tags.pop('rpc.params', None)
+            self.data["rpc"]["baggage"] = span.tags.pop('rpc.baggage', None)
+            self.data["rpc"]["error"] = span.tags.pop('rpc.error', None)
+
+        elif span.operation_name == "sqlalchemy":
+            self.data["sqlalchemy"]["sql"] = span.tags.pop('sqlalchemy.sql', None)
+            self.data["sqlalchemy"]["eng"] = span.tags.pop('sqlalchemy.eng', None)
+            self.data["sqlalchemy"]["url"] = span.tags.pop('sqlalchemy.url', None)
+            self.data["sqlalchemy"]["err"] = span.tags.pop('sqlalchemy.err', None)
+
+        elif span.operation_name == "mysql":
+            self.data["mysql"]["host"] = span.tags.pop('host', None)
+            self.data["mysql"]["port"] = span.tags.pop('port', None)
+            self.data["mysql"]["db"] = span.tags.pop(ot_tags.DATABASE_INSTANCE, None)
+            self.data["mysql"]["user"] = span.tags.pop(ot_tags.DATABASE_USER, None)
+            self.data["mysql"]["stmt"] = span.tags.pop(ot_tags.DATABASE_STATEMENT, None)
+            self.data["mysql"]["error"] = span.tags.pop('mysql.error', None)
+
+        elif span.operation_name == "postgres":
+            self.data["pg"]["host"] = span.tags.pop('host', None)
+            self.data["pg"]["port"] = span.tags.pop('port', None)
+            self.data["pg"]["db"] = span.tags.pop(ot_tags.DATABASE_INSTANCE, None)
+            self.data["pg"]["user"] = span.tags.pop(ot_tags.DATABASE_USER, None)
+            self.data["pg"]["stmt"] = span.tags.pop(ot_tags.DATABASE_STATEMENT, None)
+            self.data["pg"]["error"] = span.tags.pop('pg.error', None)
+
+        elif span.operation_name == "mongo":
+            service = "%s:%s" % (span.tags.pop('host', None), span.tags.pop('port', None))
+            namespace = "%s.%s" % (span.tags.pop('db', "?"), span.tags.pop('collection', "?"))
+
+            self.data["mongo"]["service"] = service
+            self.data["mongo"]["namespace"] = namespace
+            self.data["mongo"]["command"] = span.tags.pop('command', None)
+            self.data["mongo"]["filter"] = span.tags.pop('filter', None)
+            self.data["mongo"]["json"] = span.tags.pop('json', None)
+            self.data["mongo"]["error"] = span.tags.pop('error', None)
+
+        elif span.operation_name == "log":
+            # use last special key values
+            for l in span.logs:
+                if "message" in l.key_values:
+                    self.data["log"]["message"] = l.key_values.pop("message", None)
+                if "parameters" in l.key_values:
+                    self.data["log"]["parameters"] = l.key_values.pop("parameters", None)
+        else:
+            logger.debug("SpanRecorder: Unknown exit span: %s" % span.operation_name)
+
+    def _collect_http_tags(self, span):
+        self.data["http"]["host"] = span.tags.pop("http.host", None)
+        self.data["http"]["url"] = span.tags.pop(ot_tags.HTTP_URL, None)
+        self.data["http"]["path"] = span.tags.pop("http.path", None)
+        self.data["http"]["params"] = span.tags.pop('http.params', None)
+        self.data["http"]["method"] = span.tags.pop(ot_tags.HTTP_METHOD, None)
+        self.data["http"]["status"] = span.tags.pop(ot_tags.HTTP_STATUS_CODE, None)
+        self.data["http"]["path_tpl"] = span.tags.pop("http.path_tpl", None)
+        self.data["http"]["error"] = span.tags.pop('http.error', None)
+
+        if span.operation_name == "soap":
+            self.data["soap"]["action"] = span.tags.pop('soap.action', None)
